+++ RELEASE NOTES.md
@@ -1,11 +1,14 @@
 Version 1.3 beta
 
 - SwipeView now requires ARC (see README for details)
+- Now supports animated item view resizing and screen rotation
+- No longer crashes sometimes on iOS 7
 - The scrollOffset property is now public
 - Added scrollByOffset:duration: and scrollToOffset:duration: methods
 - Calling reloadData no longer resets scroll position
 - Fixed crash on iOS 7
-- SwipeView no longer behaves strangely if there is only one item and wrap is enabled
+- No longer behaves strangely if there is only one item and wrap is enabled
+- Now conforms to -Wextra warning level
 
 Version 1.2.10